[KVCache] Support model specific KVCache planning
