+++ vllm/v1/worker/gpu_model_runner.py
@@ -6220,20 +6220,18 @@ class GPUModelRunner(
         gc.collect()
 
     def _init_minimal_kv_cache_for_profiling(self) -> None:
-        from vllm.v1.core.kv_cache_utils import (
-            get_kv_cache_config_from_groups,
-            get_kv_cache_groups,
-        )
+        from vllm.v1.core.kv_cache_utils import get_kv_cache_planner
 
         kv_cache_spec = self.get_kv_cache_spec()
-        kv_cache_groups = get_kv_cache_groups(self.vllm_config, kv_cache_spec)
+        kv_cache_planner = get_kv_cache_planner(self.vllm_config)
+        kv_cache_groups = kv_cache_planner.get_kv_cache_groups(kv_cache_spec)
         min_blocks = self.compilation_config.max_cudagraph_capture_size or 1
 
         # Temporarily change num_gpu_blocks_override to allocate a minimal KV cache
         saved_override = self.cache_config.num_gpu_blocks_override
         self.cache_config.num_gpu_blocks_override = min_blocks
-        minimal_config = get_kv_cache_config_from_groups(
-            self.vllm_config, kv_cache_groups, available_memory=0
+        minimal_config = kv_cache_planner.get_kv_cache_config_from_groups(
+            kv_cache_groups, available_memory=0
         )
         self.cache_config.num_gpu_blocks_override = saved_override
